- python tests fixes - TicketType enum to str conversion
